app/requirements.txt
# apt install libpcsclite-dev swig pcscd xsel libgstreamer1.0-dev libsdl2-dev libsdl2-image-dev libsdl2-mixer-dev libsdl2-net-dev libsdl2-ttf-dev libpython3-dev
pyscard
pillow
requests
kivy
imageio
